refactor : like 로직 단일화.
